requirements.txt
chromadb==1.5.7
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-genai==1.73.1
googleapis-common-protos==1.74.0
groq==0.37.1
langchain-core==1.3.0
langchain-groq==1.1.2
pdfplumber==0.11.9
python-dotenv==1.2.2
streamlit==1.56.0